Module dilithium::poly::lvl2

Constants

UNIFORM_ETA_NBLOCKS 🔒

UNIFORM_GAMMA1_NBLOCKS 🔒

Functions

challenge

Implementation of H. Samples polynomial with TAU nonzero coefficients in {-1,1} using the output stream of SHAKE256(seed).

decompose

For all coefficients c of the input polynomial, compute high and low bits c0, c1 such c mod Q = c1*ALPHA + c0 with -ALPHA/2 < c0 <= ALPHA/2 except c1 = (Q-1)/ALPHA where we set c1 = 0 and -ALPHA/2 <= c0 = c mod Q - Q < 0. Assumes coefficients to be standard representatives.

eta_pack

Bit-pack polynomial with coefficients in [-ETA,ETA]. Input coefficients are assumed to lie in [Q-ETA,Q+ETA].

eta_unpack

Unpack polynomial with coefficients in [-ETA,ETA].

make_hint

Compute hint polynomial, the coefficients of which indicate whether the low bits of the corresponding coefficient of the input polynomial overflow into the high bits.

rej_eta

Sample uniformly random coefficients in [-ETA, ETA] by performing rejection sampling using array of random bytes.

uniform_eta

Sample polynomial with uniformly random coefficients in [-ETA,ETA] by performing rejection sampling using the output stream from SHAKE256(seed|nonce).

uniform_gamma1

Sample polynomial with uniformly random coefficients in [-(GAMMA1 - 1), GAMMA1 - 1] by performing rejection sampling on output stream of SHAKE256(seed|nonce).

use_hint

Use hint polynomial to correct the high bits of a polynomial.

use_hint_ip

Use hint polynomial to correct the high bits of a polynomial in place.

w1_pack

Bit-pack polynomial w1 with coefficients in [0, 15]. Input coefficients are assumed to be standard representatives.

z_pack

Bit-pack polynomial z with coefficients in [-(GAMMA1 - 1), GAMMA1 - 1]. Input coefficients are assumed to be standard representatives.*

z_unpack

Unpack polynomial z with coefficients in [-(GAMMA1 - 1), GAMMA1 - 1]. Output coefficients are standard representatives.